created a error component and an error slice
