+++ src/Pages/Accommodation/Accommodations.jsx
@@ -1,6 +1,5 @@
 import AllAccommodations from './AllAccommodations';
 import BreadCrumbs from '../../Global/BreadCrumbs';
-import Error from '../../Layout/Error';
 
 function Accommodations() {
   const breadcrumb = [
@@ -10,7 +9,6 @@ function Accommodations() {
   return (
     <div>
       <div className="mt-40 max-w-7xl mx-auto">
-        <Error />
         <BreadCrumbs breadcrumb={breadcrumb} />
         <AllAccommodations />
       </div>
